requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a1659+g700dbb5
idna==2.8
multidict==4.5.2
pycparser==2.19
PyNaCl==1.0.1
six==1.12.0
websockets==6.0
yarl==1.3.0
